Added more benchmarks

+++ asm/run.sh
@@ -3,11 +3,15 @@ SCRIPTPATH="$( cd "$(dirname "$0")" ; pwd -P )"
 cd "$SCRIPTPATH"
 
 for dir in *; do
-	[ -d "$SCRIPTPATH/$dir" ] &&
-	cd "$SCRIPTPATH/$dir" &&
-	nasm -f elf64 -o $dir.o $dir.s &&
-	ld --strip-all -n -o $dir $dir.o &&
-	rm $dir.o &&
-	echo "asm/$dir" &&
-	time ./$dir
+	if [ -d "$SCRIPTPATH/$dir" ]
+	then
+		cd "$SCRIPTPATH/$dir" &&
+		nasm -f elf64 -o $dir.o $dir.s &&
+		ld --strip-all -n -o $dir $dir.o &&
+		rm $dir.o &&
+		echo &&
+		echo "asm/$dir" &&
+		time ./$dir
+		#echo $?
+	fi
 done
